Extract panel to panels/ dir
