+++ src/panels/BeaconSummaryPanel.tsx
@@ -7,6 +7,7 @@ import {
 } from "@material-ui/core";
 import ErrorOutlineIcon from "@material-ui/icons/ErrorOutline";
 import React, { FunctionComponent, useEffect, useState } from "react";
+import { ViewPanel } from "../components/dataPanel/ViewPanel";
 import { IBeacon } from "../entities/IBeacon";
 import { IBeaconsGateway } from "../gateways/IBeaconsGateway";
 import {
@@ -14,7 +15,6 @@ import {
   formatOwners,
   formatUses,
 } from "../useCases/mcaWritingStyleFormatter";
-import { ViewPanel } from "./dataPanel/ViewPanel";
 
 interface IBeaconSummaryProps {
   beaconsGateway: IBeaconsGateway;
